[ALDN-93] Add a script to replace /Users with nfs mount
This relies on the user being able to export their host machine's `/Users` directory (or equivalent)
as an nfs mount point at `/Users`. This should allow us to have a consistent aladdin interface to
that directory and not have to deal with Cygwin-specific directory paths, like `/c/Users`. Not to
mention that we'll get the performance and permission-correctness benefits of `nfs` over `vboxsf`
mounts.

[ALDN-93](https://fivestars.atlassian.net/browse/ALDN-93)