Copy the ~/.ssh directory into minikube for Cygwin users
Since the aladdin container runs as root, we need to somehow provide the ssh
directory/credentials to the container as the root user. Cygwin/Windows users
have to do an end run around the global NFS mount, due to strict enforcement
of the file ownership and permissions there.

This adds a step to the aladdin script to `minikube mount` the host's `~/.ssh`
directory and copy then adjust the contents of the directory before volume-
mounting it into the aladdin container. This work is only performed for Cygwin
users.