[ALDN-126] Don't fail to enumerate pods when some are Failed (#28)
* [ALDN-126] Don't fail to enumerate pods when some are Failed

Under some circumstances (at least when a pod has been evicted),
pod.status.container_statuses will be None and not enumerable.  We don't
check for this, so if any pod is in such a state, running `aladdin
connect` without arguments will fail and not allow specifying a
deployment to connect to.  This PR catches that case.

[ALDN-126](https://fivestars.atlassian.net/browse/ALDN-126)

* Add comment