[ALDN-157] Fix regex for copying kube config into aladdin container (#40)
The regex we use to replace e.g. /Users/ptr/.minikube with /root/.minikube
was overly greedy, swallowing up path components that it shouldn't.  It
wasn't a problem while client.key and client.crt were in the top level of
.minikube/, but more recent minikube puts them in
.minikube/profiles/minikube, where aladdin now can't find them.

This commit updates the regex to just replace the path components up
to .minikube/.

[ALDN-157](https://fivestars.atlassian.net/browse/ALDN-157)