Add support for commands-base:2.0.0
The `commands-base:2.0.0` image moves the `command.py` script to a new location and name:
`/usr/local/bin/aladdin_cmd`. It also assumes resposibility for ensuring it's executable so
aladdin can just invoke the script or program. This is to allow developers to mount their host files
into their command containers at `/code` without stomping on this necessary script.

Since this is an interface breaking change, we are also adding a new `lamp.json` config value and
structure to support it.

```
{
    "name": "...",
    "build_docker": "...",
    "helm_chart": "...",
    "docker_images": [...],
    "commands": {
        "cmd": {
            "version": 2
        }
    }
}
```

Projects that wish to move to the new `commands-base:2.0.0` image should update their commands`
Dockerfiles to derive from this image and then set the `commands.cmd.version` to `2`. This will
instruct aladdin to invoke the command script from the new location. Additionally, if one wishes to
use an alternative implementation of the command script provided by `commands-base`, one only needs
to COPY their implementation to `/usr/local/bin/aladdin_cmd` and set it as executable.

[$JIRA](https://fivestars.atlassian.net/browse/$JIRA)