Add support for commands-base:2.0.0 (#44)
The `commands-base:2.0.0` image moves the `command.py` script to a new location and name:
`/usr/local/bin/aladdin_command`. It also assumes responsibility for ensuring that it is executable so
aladdin can just invoke the script or program. This is to allow developers to mount their host files
into their command containers at `/code` without stomping on this necessary script.

Projects that wish to move to the new `commands-base:2.0.0` image should update their commands'
Dockerfiles to derive from this image. Aladdin will check for the existence of an executable named
`aladdin_command` on the `PATH` before falling back to the previous behavior.

If one wishes to use an alternative implementation of the command script provided by
`commands-base:2.0.0`, one only needs to create an executable named `aladdin_command` on their
image's `PATH` that accepts the same parameter's as the default implementation in
`commands-base:2.0.0`. (from #26)

See fivestars-os/commands-base#3 for further details.