Disable NFS mounts by default and optimize copying of SSH keys into minikube (#48)
Cygwin users require that their `~/.ssh` directory be copied into the minikube VM to avoid
file permission issues. We only need to copy the SSH keys once, not every time aladdin is
invoked, so we now move this to the `init` procedure. The copying process has also been updated to be more resilient to repeated invocations.

Since NFS usage is an optimization and can cause issues of its own, it's now behind a default-disabled aladdin config, `nfs`. Anyone depending on it will need to `aladdin config set nfs 1` to re-enable this behavior.