Add git pre-push checks to prevent secret leaks (#54)
This adds [trufflehog](https://github.com/dxa4481/truffleHog) and [AWS git-secrets](https://github.com/awslabs/git-secrets) to our git pre-push hooks. The checks will
examine the entire repository, not just what is being pushed. If this proves to be too
time-consuming, we can revisit this in the future. But for now we are erring on the
side of safety.

This hook assumes you have installed [git-hooks](https://github.com/fivestars/git-hooks).