[ALDN-190] Add DNS name information to helm values for use in templates
In some situations it is helpful for the templates to be able to use the domain name the service
will be reachable at. One use case is the upcoming Ambassador `Mapping` CRDs we will be using for
edge-authenticated API calls to our services.

Additionally, we need to add not just the namespaced service hostname, but also one for the new
concept of the "global" namespace. By default, this will be the "default" namespace but can be
overriden by providing the `service_global_namespace` value in the cluster's aladdin config. This
global hostname is useful for identifying services that are not specific to a certain namespace
(like all of our services in the KDEV cluster), but instead exist outside of those and are common
to all services in the cluster.

Both of these hostname values are calculateable in the same manner as the existing
`service.certificateArn` value. If the `service_dns_suffix` value is defined in the cluster's
aladdin config, it will be used for both the namespace hostname as well as the global hostname.

[ALDN-190](https://fivestars.atlassian.net/browse/ALDN-190)