[ALDN-193] Support deploying a specific chart in a project
In some situations, a project may need to maintain two or more charts. `aladdin start`, `aladdin
publish` and `aladdin deploy` all currently support an all-or-nothing approach to their respective
operations. We should be able to specify which chart we wish to deploy.

This adds the `--chart` option to the `aladdin start` and `aladdin deploy` commands. The default
behavior for `start` is to start all charts if not otherwise specified. For `deploy`, it will
default to using the project's name (which prior to this was a hard-coded assumption). `deploy` will
only ever deploy one chart at a time. Multiple invocations are required to deploy multiple charts.

`aladdin publish` will now push each chart individually to S3 and `aladdin deploy` pulls all charts
down before deciding which one to publish according to the `--chart` value.

- [ ] Update docs

[ALDN-193](https://fivestars.atlassian.net/browse/ALDN-193)