[ALDN-193] Support deploying a specific chart in a project (#59)
In some situations, a project may need to maintain two or more charts. Commands like
`aladdin start`, `aladdin publish` and `aladdin deploy` all currently support an all-or-nothing
approach to their respective operations. We should be able to specify which chart we wish to deploy.

This adds the `--chart` option to the `deploy', 'restart', 'rollback', 'start', 'stop', and
'undeploy' commands. The default behavior for `start`, `stop` and `restart` is to perform the
operation on all charts if not otherwise specified. For the others, it will default to using the
project's name (which prior to this was a hard-coded assumption) and will only ever use a single
chart per invocation. Multiple invocations are required to deploy multiple charts.

This also adds the `.Values.project.name` and `.Values.deploy.namespace` values to the helm
templating process during the `start`, `restart` and `deploy` commands.

`aladdin publish` will now push all charts individually to S3 (there is no need to specify the chart
to publish) and `aladdin deploy` now pulls all charts down before deciding which one to publish
according to the `--chart` value.

[ALDN-193](https://fivestars.atlassian.net/browse/ALDN-193)