[ALDN-205] Use with_aladdin_perms_wrapper for direct aladdin bash commands
The recently added direct `aladdin bash` command invocation functionality did not provide support
for adopting the intended permissions available to other aladdin commands and tools like `kubectl`.

This wraps the provided command with `with_aladdin_perms_wrapper` when invoking it directly in a
bash shell, thus providing the requisite permissions for the duration of the command.

[ALDN-205](https://fivestars.atlassian.net/browse/ALDN-205)